start of making user icon element, moved social icons
