+++ src/components/ui/socials/socials.jsx
@@ -1,7 +1,7 @@
 import React from "react"
-import TelegramIcon from "./telegram-icon.svg"
-import DiscordIcon from "./discord-icon.svg"
-import TwitterIcon from "./twitter-icon.svg"
+import TelegramIcon from "../../../assets/social-icons/telegram-icon.svg"
+import DiscordIcon from "../../../assets/social-icons/discord-icon.svg"
+import TwitterIcon from "../../../assets/social-icons/twitter-icon.svg"
 import "./style.css"
 
 export default function Socials() {
